Update Beautiful Matrix.cpp

+++ 600/Beautiful Matrix.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-  int mat[5][5];
+ ll mat[5][5];
   int row,col;
   for(int i=0;i<5;i++)
   {
